// lib/src/io/bni/bni-building.d.ts
import { Vector2 } from "../../vector2";
import { Orientation } from '../../enums/orientation';
export declare class BniBuilding {
    offset: Vector2;
    buildingdef: string;
    orientation: Orientation;
    flags: number;
    selected_elements: number[];
}
//# sourceMappingURL=bni-building.d.ts.map